Moving utils to mgt-core. Fixing imports
