+++ src/mgt-core/components/templatedComponent.ts
@@ -6,9 +6,9 @@
  */
 
 import { html, PropertyValues } from 'lit-element';
-import { MgtBaseComponent } from '..';
 import { TemplateHelper } from '../utils/TemplateHelper';
 import { equals } from '../utils/Utils';
+import { MgtBaseComponent } from './baseComponent';
 
 /**
  * Lookup for rendered component templates and contexts by slot name.
