+++ src/mgt-core/index.ts
@@ -7,15 +7,18 @@
 
 export * from './Graph';
 export * from './IGraph';
+export * from './Providers';
+export * from './ProxyGraph';
 export * from './components/baseComponent';
 export * from './components/templatedComponent';
 export * from './components/providers/baseProvider';
 export * from './providers/IProvider';
-export * from './providers/Providers';
 export * from './utils/Batch';
 export * from './utils/ComponentMiddlewareOptions';
+export * from './utils/CustomHeaderMiddleware';
 export * from './utils/EventDispatcher';
 export * from './utils/GraphHelpers';
 export * from './utils/SdkVersionMiddleware';
+export * from './utils/SvgHelper';
 export * from './utils/TemplateHelper';
 export * from './utils/Utils';
