+++ src/mgt/components/mgt-person-card/mgt-person-card.ts
@@ -7,10 +7,17 @@
 
 import * as MicrosoftGraph from '@microsoft/microsoft-graph-types';
 import { customElement, html, property, TemplateResult } from 'lit-element';
-import { getEmailFromGraphEntity, MgtTemplatedComponent, Providers, ProviderState } from '../../../mgt-core';
-import { getSvg, SvgIcon } from '../../utils/SvgHelper';
+import {
+  getEmailFromGraphEntity,
+  getSvg,
+  MgtTemplatedComponent,
+  Providers,
+  ProviderState,
+  SvgIcon
+} from '../../../mgt-core';
 import { MgtPerson } from '../mgt-person/mgt-person';
 import { styles } from './mgt-person-card-css';
+
 /**
  * Web Component used to show detailed data for a person in the
  * Microsoft Graph
