+++ src/mgt/components/mgt-person/mgt-person.ts
@@ -11,9 +11,32 @@ import { classMap } from 'lit-html/directives/class-map';
 import { getEmailFromGraphEntity, MgtTemplatedComponent, Providers, ProviderState } from '../../../mgt-core';
 import '../../styles/fabric-icon-font';
 import { MgtPersonCard } from '../mgt-person-card/mgt-person-card';
-import { PersonCardInteraction } from '../PersonCardInteraction';
 import '../sub-components/mgt-flyout/mgt-flyout';
 import { styles } from './mgt-person-css';
+
+/**
+ * Defines how a person card is shown when a user interacts with
+ * a person component
+ *
+ * @export
+ * @enum {number}
+ */
+export enum PersonCardInteraction {
+  /**
+   * Don't show person card
+   */
+  none,
+
+  /**
+   * Show person card on hover
+   */
+  hover,
+
+  /**
+   * Show person card on click
+   */
+  click
+}
 
 /**
  * The person component is used to display a person or contact by using their photo, name, and/or email address.
